+++ docassemble_webapp/docassemble/webapp/users/models.py
@@ -25,6 +25,7 @@ class UserModel(db.Model, UserMixin):
     pypi_username = db.Column(db.String(255), nullable=True)
     pypi_password = db.Column(db.String(255), nullable=True)
     modified_at = db.Column(db.DateTime())
+    last_login = db.Column(db.DateTime())
     #email_is_phone_number = db.Column(db.Boolean(), nullable=True, server_default='0')
 
 class UserAuthModel(db.Model, UserMixin):